refactor: 로또리스트 관리하는 객체 생성
